common/rl-manager/gym_cloudsimplus/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gym_cloudsimplus"
version = "0.1.0"
description = "CloudSim Plus gRPC bridge for Stable Baselines3"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.80.0",
    "protobuf>=7.34.1",
]
